chore: finish the course

+++ exercises/23_conversions/using_as.rs
@@ -3,22 +3,22 @@
 // with renaming imports.
 
 fn average(values: &[f64]) -> f64 {
-    let total = values.iter().sum::<f64>();
-    // TODO: Make a conversion before dividing.
-    total / values.len()
+  let total = values.iter().sum::<f64>();
+  // TODO: Make a conversion before dividing.
+  total / values.len() as f64
 }
 
 fn main() {
-    let values = [3.5, 0.3, 13.0, 11.7];
-    println!("{}", average(&values));
+  let values = [3.5, 0.3, 13.0, 11.7];
+  println!("{}", average(&values));
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+  use super::*;
 
-    #[test]
-    fn returns_proper_type_and_value() {
-        assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
-    }
+  #[test]
+  fn returns_proper_type_and_value() {
+    assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
+  }
 }
